requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
django-celery-beat
certifi==2025.6.15
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.9
Django==5.2.3
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.18.0
gunicorn==23.0.0
inflection==0.5.1
kombu==5.5.4
packaging==25.0
pillow==11.2.1
pipenv==2025.0.3
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
pytest
pytest-django
pytest-factoryboy    # if you want to use factories
factory-boy
